Fixed a serious bug in save_image method when the saving data's type is int or uint and so on while the first import data is other type.

+++ froi/gui/component/regularroifromcsvfiledialog.py
@@ -105,13 +105,12 @@ class RegularROIFromCSVFileDialog(QDialog):
         header = self._model.data(self._model.currentIndex(), Qt.UserRole + 11)
         image_affine = self._model.get_affine()
         data = self._model.data(self._model.currentIndex(), Qt.UserRole + 6)
-        new_data = np.zeros_like(data)
+        new_data = np.zeros_like(data).astype(np.uint32)
         try:
             coord_list, radius_list, id_list = get_cord_from_file(header, cord_filepath, image_affine)
         except ValueError, error_info:
             QMessageBox.critical(self, 'Please check the cordinate in the file.', str(error_info))
             return
-
         for idx in range(len(coord_list)):
             new_data = roi_generater(new_data, coord_list[idx][0],
                                     coord_list[idx][1], coord_list[idx][2],
